[WARP-SPEED] SIC-2871 simplify match result
(cherry picked from commit fe2960a6741b57c20213de9cfcf59c5762f24101)

+++ plugin/starburst-warp-speed/starburst-warp-speed-core/src/main/java/io/trino/plugin/warp/storage/lucene/LuceneMatcher.java
@@ -137,7 +137,7 @@ public class LuceneMatcher
             WarpInputDirectory warpInputDirectory = new WarpInputDirectory(luceneIndexReader,
                     storageEngineConstants,
                     lucenePageCacheStats,
-                    (int) indexUniqueIdInRowGroup,
+                    indexUniqueIdInRowGroup,
                     FILE_PREFIX,
                     chunkState.filesLength());
             IndexReader reader = DirectoryReader.open(warpInputDirectory);
